#53 [feat] : GalleryUploadViewModel 추가
